requirements.txt
scikit-learn==0.23.0
confluent-kafka==1.7.0
slackclient==2.9-3
python-dotenv==0.17.1